Experimental solution examples used from the TF 2.0 teutorials
DO NOT USE FOR PRODUCTION: AS IS existing model is a very low performing accuracy to highlight that the original vanilla dataset columns without additional data analytics and the default linear modeling used in the solution only has a 60-68% accuracy. Additional data feature extrapolation and carving required along with more model tuning is required before putting this into production with a .95 or greater accuracy.